This change alters Open Issue B.1 to use path in place of simple name to identify a mount point within a module.
It also makes extension less mount points an *option* vs the only
option.  Many in the WG have already voiced their view on the importance
of the use of the extension when designing modules that contain mount
point.  And removing the ability to include the extension would against
this apparent consensus position.